// frontend/src/generated/k9b-api/models/ExecuteNextCheckRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * k9b API
 * Machine-readable API contract for k9b incidents, snapshots, diagnosis loop, auth, runtime status, and diagnostics endpoints.
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * Next-check execution request
 * @export
 * @interface ExecuteNextCheckRequest
 */
export interface ExecuteNextCheckRequest {
    /**
     * Unique candidate identifier
     * @type {string}
     * @memberof ExecuteNextCheckRequest
     */
    candidateId?: string;
    /**
     * Zero-based index of candidate in plan
     * @type {number}
     * @memberof ExecuteNextCheckRequest
     */
    candidateIndex?: number;
    /**
     * Target cluster label
     * @type {string}
     * @memberof ExecuteNextCheckRequest
     */
    clusterLabel: string;
    /**
     * Explicit plan artifact path
     * @type {string}
     * @memberof ExecuteNextCheckRequest
     */
    planArtifactPath?: string;
}

/**
 * Check if a given object implements the ExecuteNextCheckRequest interface.
 */
export function instanceOfExecuteNextCheckRequest(value: object): value is ExecuteNextCheckRequest {
    if (!('clusterLabel' in value) || value['clusterLabel'] === undefined) return false;
    return true;
}

export function ExecuteNextCheckRequestFromJSON(json: any): ExecuteNextCheckRequest {
    return ExecuteNextCheckRequestFromJSONTyped(json, false);
}

export function ExecuteNextCheckRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): ExecuteNextCheckRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'candidateId': json['candidateId'] == null ? undefined : json['candidateId'],
        'candidateIndex': json['candidateIndex'] == null ? undefined : json['candidateIndex'],
        'clusterLabel': json['clusterLabel'],
        'planArtifactPath': json['planArtifactPath'] == null ? undefined : json['planArtifactPath'],
    };
}

export function ExecuteNextCheckRequestToJSON(json: any): ExecuteNextCheckRequest {
    return ExecuteNextCheckRequestToJSONTyped(json, false);
}

export function ExecuteNextCheckRequestToJSONTyped(value?: ExecuteNextCheckRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'candidateId': value['candidateId'],
        'candidateIndex': value['candidateIndex'],
        'clusterLabel': value['clusterLabel'],
        'planArtifactPath': value['planArtifactPath'],
    };
}